+++ hostfunctions/hostfns.go
@@ -28,6 +28,8 @@ func Instantiate(ctx context.Context, runtime *wazero.Runtime) error {
 	b.NewFunctionBuilder().WithFunc(hostComputeEmbedding).Export("computeEmbedding")
 	b.NewFunctionBuilder().WithFunc(hostInvokeTextGenerator).Export("invokeTextGenerator")
 	b.NewFunctionBuilder().WithFunc(hostFetch).Export("httpFetch")
+	b.NewFunctionBuilder().WithFunc(hostLookupModel).Export("lookupModel")
+	b.NewFunctionBuilder().WithFunc(hostInvokeModel).Export("invokeModel")
 
 	_, err := b.Instantiate(ctx)
 	if err != nil {