+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## UNRELEASED
+## 2024-06-21 - Version 0.9.0
 
 - Add nullable check in ReadString [#228](https://github.com/hypermodeAI/runtime/pull/228)
 - Lowercase model name before invoking for hypermode hosted models [#221](https://github.com/gohypermode/runtime/pull/221)